Restored focus on up next view after refreshing up next events
Before the fix: the focus on the CalendarEventListItemViewJelly will be
lost after the up next view is refreshed every 15 seconds.
This issue is fixed by saving the focused view before updating the
events. Then if the focused view is a child view of the up next content
view, restore the focus on it after the refresh.

Bug: b:286596205
Change-Id: I81c41b0346db2de068e5b0499187d6c428277c71
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4629589
Reviewed-by: Jiaming Cheng <jiamingc@chromium.org>
Commit-Queue: Sylvie Liu <sylvieliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1160812}