[webauthn] Synced GPM passkeys on conditional UI
Display synced GPM passkeys on webauthn's conditional UI. Credentials
are shown for users who have opted in to syncing them. Credentials may
also show up on the platform authenticator credential picker. This is
okay as we'll update that UI surface before launching the feature.

For now, credentials are rendered with the subtext "Use lock screen". A
follow-up will change this to display the device name.

Tapping a GPM passkey dispatches to a sync paired phone. At the moment,
this is the first applicable phone on the list. A follow-up will have
Chrome remember the last used phone instead.

Finally, with the feature enabled IsConditionalMediationAvailable will
always return true on all desktop platforms since they are all capable
of showing GPM passkeys.

This feature is guarded by the WebAuthenticationListSyncedPasskeys flag
and requires SyncWebauthnCredentials to be enabled as well.

Bug: 1428655
Change-Id: I7297935f905f443eae4b4b0621128c2bc47f0966
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4621635
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Commit-Queue: Nina Satragno <nsatragno@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1160857}