Adjust snapped window bounds on workspace event
This CL fixes the snapped window bounds in response to workspace events.
When any workspace event happens, the snapped window bounds need to be
adjusted, but it should keep the snap ratio instead of resetting it to
the *default* snapped bounds.

Example: Let’s say the workarea width is 1200dp. After resizing a
snapped window width to be like 600dp->300dp, the snap ratio is 0.25
(=300/1200). When the workarea width is changed to 800dp, the window
width should be updated to 200dp (=800*0.25) because we need to keep the
snap ratio 0.25.

TEST=ClientControlledStateTestClamshellAndTablet.ResizeSnappedWindow
BUG=b:235312364
BUG=b:207597527
BUG=b:178766080
BUG=b:280225357

Change-Id: Id0f1d31734932f84b340b44f7e070824fab6a7f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4622491
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Commit-Queue: Toshiki Kikuchi <toshikikikuchi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1161011}