[FilesBulkPinning] Enable disabled tests
Previously there was an OR condition to respect the
FeatureManagementDriveFsBulkPinning flag OR the DriveFsBUlkPinning flag
prior to M117. From M117 onwards it would only respect both of them.
This caused the tests to fail when both flags aren't supplied.
CL:4635238 removed this condition, so update the tests to include both
flags and remove the DISABLED_ prefix.

Fixed: 1456621, 1456627, 1456602
Change-Id: Ifaa41a1853587ae2e03519e3be41b9815ef113a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4633121
Reviewed-by: François Degros <fdegros@chromium.org>
Commit-Queue: François Degros <fdegros@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1161039}