[coverage] Add convenience script to run all fuzzers in parallel while preserving the resulting profdata files.
Being able to run all fuzzers, in separate processes, is crucial for
efficiently generating a coverage dashboard & tracking individual fuzzer
jobs that fail for whatever reason.

This patch is part of a series, and the overall goal of the series is to
provide a set of convenience scripts to make it simpler to generate
fuzzing coverage reports.

Change-Id: I50b1a204ec521014e687a6f6ad908cc22ec3b7cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4633629
Reviewed-by: Prakhar Asthana <pasthana@google.com>
Commit-Queue: Julia Hansbrough <flowerhack@google.com>
Cr-Commit-Position: refs/heads/main@{#1161048}