Remove NOTREACHED for infobar in translating step
It seems possible that, if the translate infobar is queued
behind another bar and the status changes before it is
displayed, that the infobar is displayed when its status
is TRANSLATE_STEP_TRANSLATING.
I could not reproduce, but there are multiple reports
showing that this happens.
This CL handles this case properly instead of NOTREACHING.

It does not seem possible to reach this state in TRANSLATE_STEP_NEVER_TRANSLATE, so keep the NOTREACHED in
that case. This can be reevaluated is the NOTREACHED is
still triggers after this CL.

Fixed: 1454237, 1421728
Change-Id: I730103f5decfb134bc5f068968fce880027e30ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4614932
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1161114}