Reland "ChromeDriver supports pipes"
This is a reland of commit 88edec05ebc57202610f3c62242a9ac312c6f1a6

The default communication mode is restored to remote-debugging-port. The code in this commit is not invoked in this mode and therefore no flakiness in the existing tests can be attributed to this commit.

Original change's description:
> ChromeDriver supports pipes
>
> This commit adds support for communication over pipes between ChromeDriver and Chrome.
> Pipes on Windows are not yet supported.
>
> Bug: chromedriver:3480
> Change-Id: Ic4d145fde3089d28f07a683867b0b38b3e8de2f2
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4566463
> Commit-Queue: Vladimir Nechaev <nechaev@chromium.org>
> Reviewed-by: Thiago Perrotta <tperrotta@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1160198}

Bug: chromedriver:3480
Change-Id: I5a7b643d93e7e277a12836fd633cdc2b9191517e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4635378
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Vladimir Nechaev <nechaev@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1161129}