[Camera Roll] Cleanup expired camera roll flags
Feature is already enabled by default and launched. This change does not
remove the flag in ash_features in order to keep that as a kill switch,
this change only removes the aspects of the flag that would expose it to
the user.

Test: manually verified Camera Roll still WAI
Fixed: 1419223
Change-Id: I49eb748d53d49a47b2338ec41731787e1cbdab85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4632929
Reviewed-by: Pu Shi <pushi@google.com>
Commit-Queue: Crisrael Lucero <crisrael@google.com>
Cr-Commit-Position: refs/heads/main@{#1161167}