Store is_autofilled attribute in Undo history
It was assumed that we always undo autofilled fields to non-autofilled
fields. This is wrong because we can trigger Autofill from an autofileld
field by focusing on the field in question.

This CL adds the is_autofilled state to the history because it is now
something we can't just assume to be false.

Bug: 1441410
Change-Id: I14af32850e0cc8dc9a435674f57c2503c8c3d00f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4628909
Commit-Queue: Jihad Hanna <jihadghanna@google.com>
Reviewed-by: Christoph Schwering <schwering@google.com>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1161185}