[ios] Use DIR_ASSETS to load the assets for embedded HTTP server
The files served by the embedded HTTP server are located with the other
assets, so use DIR_ASSETS to access to the path via PathService instead
of duplicating code to locate them inside the bundle.

Ensure that DIR_ASSETS is correctly initialized in EarlGrey tests for
the iOS WebView shell.

Bug: none
Change-Id: If83bc9cdbe16a3005354f794de1ab679f795a14c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4629750
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Federica Germinario <fedegermi@google.com>
Cr-Commit-Position: refs/heads/main@{#1161186}