[ios] Optimise storage of "pinned" status of WebState
The WebStateList used a boolean in each WebStateWrapper to store whether
the WebState is pinned, and at the same time maintain the invariant that
all pinned WebStates are at the beginning of the list.

This means that the same information is duplicated (both in the boolean
and in the position). This could lead to bugs if there is divergence of
the information.

Instead it is possible to only store the number of pinned WebStates and
to check the status for an individual WebState by comparing this index
to the counter.

This optimisation is also used by the new serialisation code to save as
described at go/bling-session-restoration.

Bug: 1383087
Change-Id: I678391c1fae97b40883454121bb638fd389be525
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4628252
Reviewed-by: Petro Akzhygitov <pakzhygitov@google.com>
Reviewed-by: Asami Doi <asamidoi@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1161193}