[iOS] Fix crash in loadSignOutAndTurnOffSyncSection for kSignedOut.
In case of SyncSettingsAccountState::kSignedOut, syncDataTypeSection is
not found, and thus trying to insert the sign out section below it
crashes.

This CL skips the loadSignOutAndTurnOffSyncSection for kSignedOut case.
Note: the SyncSettingsAccountState::kSignedOut is a temp state as the UI
is signing out. So it does not actually matter what we load here.

Bug: 1456446
Change-Id: I11455e46865c07339763e4f3e7036f68cc5a396a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4633299
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Commit-Queue: Jood Hajeer <jood@google.com>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1161198}