Re-enable Flake Endorser on CQ Win
This change re-enables Flake Endorser for Win. This is part 1 of 3.

Flake Endorser was previously disabled on these builders due to a high
false positive rate, impacting users as it was repeating tests that
were unrelated to their patchset.

ResultDB Test Results optionally allow tests to upload a file path
as part of its metadata, associating a specific test to a file.
Most of Chrome have this being set.

In crrev/c/4628558, Flake Endorser utilizes that by checking that new
tests with a file path defined have that file path being modified
as part of the patchset. If not, it's skipped. If file path is not
defined, it's included, but re-verified through the TestHistory RPC call.

Bug: 1414157, 1434778
Change-Id: I9e350e0099514297a32bb0277f4d008055c37953
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4633631
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Wanda Mora <morawand@google.com>
Cr-Commit-Position: refs/heads/main@{#1161227}