Replace NOTREACHED() with histograms
This CL replaces several NOTREACHED(); calls in
AutofillWebDataBackendImpl which are handled gracefully with a UMA
histogram to monitor their prevalence and not trigger crash dumps.
It also adds histograms to several places where NOTREACHED() is
not called to collect a baseline.

Bug: 1454931,1424298,1435439,1430313,1441361
Change-Id: I8e6f68fbbcf0eb2f3bba1ce4d4f256f3fb30a5c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4614553
Reviewed-by: Peter Kotwicz <pkotwicz@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Dominic Battre <battre@chromium.org>
Reviewed-by: Caitlin Fischer <caitlinfischer@google.com>
Cr-Commit-Position: refs/heads/main@{#1161231}