Add extensions to SecurePaymentRequestData.
Before this change extensions set on the SecurePaymentConfirmation data
of a PaymentRequest would have no effect.

This change adds the extensions member to SecurePaymentRequestData and
set on the underlying WebAuth PublicKeyCredentialRequestOptions.
SecurePaymentRequestData is the data member of a PaymentRequest when the
payment method is set to "secure-payment-confirmation".

Now the extensions are handled the same as get credential calls into
WebAuthn.  No change in behaviour is expected when extensions are not
set.

BUG=1365339

Change-Id: I5cdcc14c7e1c87e56c46996367af7e7f1a1a9ab4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4608620
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Commit-Queue: Slobodan Pejic <slobodan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1161236}