Move IsMinimumAddress to a separate file
IsMinimumAddress was used at the time of the profile import from form
submission.
It will also be used to check if an existing profile can be migrated
from the settings to Google Account. This CL is in preparation for
that change.

Bug: 1407666
Change-Id: I088cdefc3f2a14458fd60eb66e4f43a4924638dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4629313
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Vidhan Jain <vidhanj@google.com>
Cr-Commit-Position: refs/heads/main@{#1161238}