printing: Fix primary user profile access
CupsProxyServiceManager is KeyedService and is created during
profile loading. It creates CupsProxyServiceDelegateImpl when
CupsProxyClient thinks underlying dbus service is available.
CupsProxyServiceDelegateImpl reaches for primary user profile
in it constructor. There is a race between profile finish
loading and dbus service availability. When dbus availability
is signaled earlier than profile loading, DumpWithoutCrashing in
ProfileManager::GetPrimaryUserProfile would be hit.

Since CupsProxyServiceManager is a KeyedService and knows the
profile, we don't have to reach out to ProfileManager to get it.
This CL does the following:
- Create CupsProxyServiceManager only for the primary user profile;
- Pass down Profile explicitly to CupsProxyServiceDelegateImpl;

Bug: 1424583
Change-Id: I6033318eff3d76126fceaeb901b57f8794c723e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4634497
Reviewed-by: Paul Moy <pmoy@chromium.org>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1161308}