[Thumbnail Placeholder] Fix matrix sometimes being wrong
resizeIconDrawable is called before layout in onMeasure and earlier.
If the view is not laid out this can result in the matrix being wrong
and the UI not drawing. Fix this by measured values instead when
computing the matrix.

Fixed: 1457218
Change-Id: I4acbdb290153e8c9ed2b66505e9ee11b9caa2f48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4637202
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Fred Mello <fredmello@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1161320}