[Auto-disable Accessibility] Fix approach to recommended timeout for accessibility
This CL fixes a bug in the implementation of recommended timeouts for
accessibility.

With a previous CL (crrev.com/c/4583365) we updated AccessibilityState
so clients can call through that rather than having a hard dependency
on the AccessibilityManager. This CL fixes a bug in that implementation
and more closely matches the native approach.

With the previous implementation, we assumed a linear multiplier
could be applied to input params to get a recommended timeout.
The output of the method is not linear, since there is a minimum
time that the Android FW will return. For example, providing a
timeout of 100ms or 200ms both return a recommended timeout of
10,000ms from the FW. In the first implementation, we tried to
keep the method simple for clients and return a multiplier, so a
client can simply do timeout*AccessibilityState#getRecommend...
Behind the scenes the AccessibilityState would calculate the
multiplier to apply for a timeout. However, given the example
above, the 100ms would be multiplied by 100, but 200ms by 50, so
the results were inconsistent and incorrect. With this CL we
change the AccessibilityState#getRecommendedTimeout method to
match the FW impl. Now a client must pass a timeout, and the method
will return the recommended timeout. The AccessibilityState now
acts as a pass-through to AccessibilityManager so that clients do
not need a hard dependency on the FW.



AX-Relnotes: N/A
Bug: 1455299
Change-Id: Ib3786654688d7011f54ec7e93a9767d30d7820ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4621338
Reviewed-by: Amanda Lin Dietz <aldietz@google.com>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Lijin Shen <lazzzis@google.com>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1161337}