[Messages] Update a11y message of secondary button.
The content description of secondary button will be in the format of
"More options {the title}".

If the client has set a content description for that icon, we
we won't update that; otherwise, we will update its content
description according to the title or the content description of
the title.

Bug: 1364378
Change-Id: I4bf4e062c9029503b5e4e85fab14e1e0a7957c34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4633849
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Lijin Shen <lazzzis@google.com>
Reviewed-by: Aishwarya Rajesh <aishwaryarj@google.com>
Cr-Commit-Position: refs/heads/main@{#1161381}