shimless: Add compliance check UI
This CL adds a stub UI for the compliance check implementation.
In the next CL, I will update the backend + mojo files to support
retrieving the actual feature level and use that to setup the UI.

Screenshots:
- Flag enabled, feature level unknown:
  https://screenshot.googleplex.com/vBw2FkH5YjdiPSW
- Flag enabled, feature level 0 (not compliant)
  https://screenshot.googleplex.com/4zGHNpEVYP7eBTP
- Flag enabled, feature level 1 (compliant)
  https://screenshot.googleplex.com/6WGaqZLkoAnM465
- Flag disabled
  https://screenshot.googleplex.com/BEKTJGPMxaF38pQ

Note: in reimaging_device_information_page_test.js, I purposefully
underscore-separated the names of the tests. It's inconsistent with the
other names, but I think it's more readable.

Test: browser_tests --gtest_filter=ShimlessRmaApp*
Bug: b/287317464
Change-Id: Ib82e710a458f3105e8ba396687fbeb28e6ca16af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4618709
Reviewed-by: Gavin Williams <gavinwill@chromium.org>
Commit-Queue: Camden Bickel <cambickel@google.com>
Cr-Commit-Position: refs/heads/main@{#1161386}