[WebAuthn] Separate Android AuthenticatorImpl cleanup from close()
Currently `close()` is called in AuthenticatorImpl.java when a
WebAuthn request completes. `close()` is overriden from the mojo
Interface class, and is also called when the connection terminates.

This change adds `cleanupRequest()` which performs the request state
cleanup, so that `close()` can be used as a clear signal that the
connection has closed.

Bug: 1457277
Change-Id: Ia20115dc15503e77ce1480fd5a568203a8fe70d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4636953
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Nina Satragno <nsatragno@chromium.org>
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1161391}