Re-enable Flake Endorser on CQ ChromeOS
This change re-enables Flake Endorser for ChromeOS. This is part 3 of 3.

Flake Endorser was previously disabled on these builders due to a high
false positive rate, impacting users as it was repeating tests that
were unrelated to their patchset.

ResultDB Test Results optionally allow tests to upload a file path
as part of its metadata, associating a specific test to a file.
Most of Chrome have this being set.

In crrev/c/4628558, Flake Endorser utilizes that by checking that new
tests with a file path defined have that file path being modified
as part of the patchset. If not, it's skipped. If file path is not
defined, it's included, but re-verified through the TestHistory RPC call.

Bug: 1414157, 1434778
Change-Id: I2fc88f5079cca96684aba1071bf4144dda13b140
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4633632
Reviewed-by: Wanda Mora <morawand@google.com>
Reviewed-by: Sven Zheng <svenzheng@chromium.org>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Cr-Commit-Position: refs/heads/main@{#1161392}