Add InterstitialInteractionMap parameter to trigger manager functions
As part of implementing interaction lists for CSBRRs, we must add a
parameter to the trigger manager's FinishCollectingThreatDetails method
that allows us to access the list of interstitial interactions. Once we
submit this, we will submit a CL that adds the actual interaction list
to CSBRRs.

Bug: 1439670
Change-Id: Ifd31cbd57f6edd5fa6b0dbb248685c5d812ca546
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4621748
Commit-Queue: Sarah Krakowiak <skrakowi@chromium.org>
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1161399}