Rename Platform.Hibernate.LoginToResumeReady to Platform.Hibernate.ResumeTime.Total
Rename the metric Platform.Hibernate.LoginToResumeReady to
Platform.Hibernate.ResumeTime.Total. The idea is to have a
namespace for resume time metrics. Also update the description
to reflect that this is time between the user logging in and
the hibernated system being restored.

As of now the LoginToResumeReady metric is not used by devices
in the field, so it can be simply renamed instead of keeping
the old name around for some time.

Bug: b:288451949
Change-Id: Ic012f0aafa618d6d91aeda8204b9d08f11d5acf1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4638349
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1161445}