Stop AppHomePageHandlerTest pinning to the taskbar
AppHomePageHandlerTest.Create* tests are pinning shortcuts to the
Windows taskbar, and leaking shortcuts on the Desktop. This CL fixes
that by using OsIntegrationTestOverrideImpl::BlockingRegistration and
SimulateDeleteShortcutsByUser.

It also adds a call to FlushShortcutTasks() after the dialog is
closed, because that creates shortcuts. This is needed because
SimulateDeleteShortcutsByUser dchecks if the expected shortcuts don't
exist.

Bug: 1454110
Change-Id: Ia1d875f750f60664f928ea2990c89772874a51c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4629880
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: David Bienvenu <davidbienvenu@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1161479}