CrOS Settings: Do not focus Network page on initial app load
When the CrOS Settings app root page is loaded (chrome://os-settings/),
the Network page should not be focused. Instead, the search box should
be the initially focused element, allowing for faster navigation for
our keyboard users. This functionality regressed after
crrev.com/c/4614412 and is fixed in this CL.

Extracted the route navigation tests to a separate test file to mitigate
timeout errors. The test is in JS instead of TS since some of the test
fixtures used are not supported in TS yet, which is out-of-scope of this
CL.

Screen recording: http://shortn/_Ownbdg3Bsb

Bug: b/285968876
Test: browser_tests --gtest_filter="OSSettingsMainPageContainer*"
Test: browser_tests --gtest_filter="OSSettingsOsSettingsUi*"
Change-Id: Ieba080b3339b21b469ec0a369dacb67bb4bcf3ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4630425
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Commit-Queue: Wes Okuhara <wesokuhara@google.com>
Cr-Commit-Position: refs/heads/main@{#1161506}