Reland "Restored focus on up next view after refreshing up next events"
This is a reland of commit 6e4ac1041abbe81f7e1966bda369c434fe0881d8

ASAN catches a heap-use-after-free crash for the previous fix. It turns
out we cannot save the focused view and check if it's contained in the
content_view_ because a different address will be allocated if a new
CalendarEventListItemViewJelly is added into the content_view_. The root
cause of the original bug is: we checked if the displayed events are the
same with the newly fetched event list, but there's a typo in the method
so that even if the events are the same, we still remove them and add
new events into the up_next_view. This CL fixed the typo so that if the
event list isn't updated, the focus should be preserved on the first
item. If the event list gets updated, the focus should be lost, which is
intended. We can reuse the same unit test that's added in the original
change to test this fix.

Original change's description:
> Restored focus on up next view after refreshing up next events
>
> Before the fix: the focus on the CalendarEventListItemViewJelly will be
> lost after the up next view is refreshed every 15 seconds.
> This issue is fixed by saving the focused view before updating the
> events. Then if the focused view is a child view of the up next content
> view, restore the focus on it after the refresh.
>
> Bug: b:286596205
> Change-Id: I81c41b0346db2de068e5b0499187d6c428277c71
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4629589
> Reviewed-by: Jiaming Cheng <jiamingc@chromium.org>
> Commit-Queue: Sylvie Liu <sylvieliu@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1160812}

Bug: b:286596205
Change-Id: Ia9072d402b4d1ebf8c080e56b44b375efaac7bef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4638446
Commit-Queue: Sylvie Liu <sylvieliu@chromium.org>
Reviewed-by: Jiaming Cheng <jiamingc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1161511}